Expose library for snapshot conversion
